refactor: move spacialSpacer, Centralizer to LayoutComponent
