+++ src/widgets/LayoutComponent/index.ts
@@ -1,4 +1,7 @@
+export { Centralizer } from "./ui/Centralizer";
 export { Header } from "./ui/Header";
 export { SearchModal } from "./ui/SearchModal";
 export { SimpleLayout } from "./ui/SimpleLayout";
 export { SpacialControl } from "./ui/SpacialControl";
+export { SpacialSpacer } from "./ui/SpacialSpacer";
+
