+++ src/widgets/LayoutComponent/ui/Centralizer.tsx
