+++ src/widgets/LayoutComponent/ui/SpacialSpacer.tsx
